Fix accidental include of upstream header
