+++ src/corecel/sys/ActionInterface.hh
@@ -236,6 +236,24 @@ class ConcreteAction : public virtual ActionInterface
 };
 
 //---------------------------------------------------------------------------//
+//! Action order/ID tuple for comparison in sorting
+struct OrderedAction
+{
+    StepActionOrder order;
+    ActionId id;
+
+    //! Ordering comparison for an action/ID
+    CELER_CONSTEXPR_FUNCTION bool operator<(OrderedAction const& other) const
+    {
+        if (this->order < other.order)
+            return true;
+        if (this->order > other.order)
+            return false;
+        return this->id < other.id;
+    }
+};
+
+//---------------------------------------------------------------------------//
 
 // Get a string corresponding to a surface type
 char const* to_cstring(StepActionOrder);
